requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
google-cloud-vision==3.9.0
anthropic==0.42.0
Pillow==11.1.0
